+++ pyobjc-core/Modules/objc/objc-class.m
@@ -2017,6 +2017,7 @@ class_setattro(PyObject* self, PyObject* name, PyObject* _Nullable value)
         return 0;
     }
 
+#if 0 /* Disabled check due to #479 */
     /* Check if there is a current attribute with the same name that
      * is an unbound selector.
      */
@@ -2033,6 +2034,7 @@ class_setattro(PyObject* self, PyObject* name, PyObject* _Nullable value)
 
         return -1;
     }
+#endif
 
     res = PyType_Type.tp_setattro(self, name, value);
     return res;